+++ example/lib/main.dart
@@ -8,12 +8,14 @@ import 'package:flutter/material.dart' hide Step;
 import 'package:flutter/services.dart';
 import 'package:survey_kit/survey_kit.dart';
 import 'package:survey_kit_audio/survey_kit_audio.dart' as ska;
+import 'package:survey_kit_lottie/survey_kit_lottie.dart';
 import 'package:survey_kit_video/survey_kit_video.dart';
 import 'package:surveykit_example/app_bar_example.dart';
 
 void main() {
   Content.registerFromJson('audio', ska.AudioContent.fromJson);
   Content.registerFromJson('video', VideoContent.fromJson);
+  Content.registerFromJson('lottie', LottieContent.fromJson);
   runApp(const MyApp());
 }