+++ packages/survey_kit_lottie/lib/src/view/widget/content/lottie_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
-import 'package:survey_kit/survey_kit.dart';
+import 'package:survey_kit_lottie/src/model/content/lottie_content.dart';
 
 class LottieWidget extends StatelessWidget {
   const LottieWidget({