+++ packages/survey_kit_lottie/lib/survey_kit_lottie.dart
@@ -1,0 +1,4 @@
+library survey_kit_lottie;
+
+export 'src/model/content/lottie_content.dart';
+export 'src/view/widget/content/lottie_widget.dart';